Rename files for clarity and consistency
